% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mpcross2int.R
\name{mpcross2int}
\alias{mpcross2int}
\title{Convert an mpcross object to an interval object}
\usage{
mpcross2int(mpcross, gen.type = "mpMarker", method = "qtl", step = 0.1,
  population = "4way", debug = FALSE)
}
\arguments{
\item{mpcross}{An \code{mpcross} object generated for example
using the \pkg{mpMap} package.}

\item{gen.type}{Founder probabilities found for intervals on the
linkage map using \code{"interval"} or markers using
\code{"marker"}.  Default is \code{"interval"}.}

\item{method}{Method for calculation of founder probabilities.
Default is \code{"qtl"} to use Hidden Markov Models in the
\pkg{qtl} package.  The alternative is \code{"3pt"} to use
probabilities found using haplotypes of length 3.}

\item{step}{Step size for calculation of founder probabilities
based on the Hidden Markov Model using integration (Trapezoidal
rule)}

\item{population}{A character vector which indicates the numer of
founders in the MAGIC population, required for the calculation of
three point probabilities.  The default is \code{"4way"} (the
other current option is \code{"8way"}).}

\item{debug}{Logical (default is \code{FALSE}) to determine if
printing is to be enabled in debugging the function}
}
\value{
An interval object (list) suitable for QTL analysis using
mpwgaim.  The list contains the following components:
\itemize{
\item \code{geno} : A list with elements named after the
corresponding names of the chromosomes or linkage groups.  Each
component of the list has potentially five components.  These are
the linkage map in \code{map}, the marker data \code{data}, marker
data for the founders in \code{founders}, and \code{dist} which
has the distances in cM for the linkage group.  The fifth
component depends on whether the analysis using \code{mpwgaim} is
based on intervals (component \code{intval} or markers (component
\code{imputed.data}).
\item \code{pheno} : A vector of four-way line names
\item \code{founders} : A vector of founder names
\item \code{nfounders} : The number of founders in the MAGIC
population
}
}
\description{
Converts an object of class \code{"mpcross"} to an object with
class \code{"interval"}.
}
\details{
This function provides the conversion of genetic data object of
class \code{"mpcross"} from the \pkg{mpMap} package to an
\code{"interval"} object that is ready to use with \pkg{mpwgaim}
QTL analysis functions.
}
\examples{
\dontrun{
## Simulated data to illustrate generating the interval object.
## We require the mpMap and qtl packages:
require(mpMap)
require(qtl)
set.seed(3)
map <- sim.map(len=rep(300,7), n.mar=rep(201,7), eq.spacing=TRUE,
               include.x=FALSE)
sim.ped <- sim.mpped(4, 1, 500, 6, 1)
## Dummy QTL to set up structure
qtl.mat <- matrix(data=c(1, 142, 0.3, -0.3, -0.3, 0.3,
                         2, 162, 0.3, -0.3, -0.3, 0.3,
                         3, 78, 0, 0, 0, 0,
                         4, 114, 0.3, -0.3, -0.3, 0.3),
                  nrow=4, ncol=6, byrow=TRUE)
## Set up the mpcross object
sim.dat <- sim.mpcross(map=map,
pedigree=sim.ped, qtl=qtl.mat, seed=5)
mpSim <- maporder(sim.dat)
## Population is four way which is the default setting.
## Interval object for analysis using intervals and Hidden Markov
## model for probability calculations
mpInterval <- mpcross2int(mpSim, gen.type="mpInterval")
## Interval object for analysis using markers and Hidden Markov
## model for probability calculations
mpMarker <- mpcross2int(mpSim, gen.type="mpMarker")
## Interval object for analysis using intervals and three-point probabilities
mpInterval <- mpcross2int(mpSim, gen.type="mpInterval", method="3pt")
## Interval object for analysis using markers and three-point probabilities
mpMarker <-mpcross2int(mpSim, gen.type="mpMarker", method="3pt")
}
}
\author{
Ari Verbyla (ari.verbyla at csiro.au)
}

