% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mpqtl.pick.R
\name{mpqtl.pick}
\alias{mpqtl.pick}
\title{QTL selection for 'mpwgaim'.}
\usage{
mpqtl.pick(asr, intervalObj, asdata, gen.type, state, verboseLev)
}
\arguments{
\item{asr}{current QTL model.}

\item{intervalObj}{interval object constructed using
'mpcross2int'.}

\item{asdata}{data frame containing the current phenotypic and
genetic information.}

\item{gen.type}{'interval' or 'marker' depending on the type of
analysis.}

\item{state}{zero-one vector indicating the current QTL (ones).}

\item{verboseLev}{controls the printed output from this function;
any value greater than zero will flag that printing of outlier
statistics is to be carried out.}
}
\value{
A list object with updated state vector, the qtl selected,
the outlier statistics for each interval or marker, and the blups
for each interval or marker.
}
\description{
Function for selecting a putative QTL in the forward selection
approach of mpwgaim.
}
\details{
This function is usually only for internal use but could be used
to provide a manual rather than automated QTL analysis for
multi-parent populations.
}
\author{
Ari Verbyla (ari.verbyla at csiro.au)
}

