% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mpwgaim-package.R
\docType{package}
\name{mpwgaim-package}
\alias{mpwgaim-package}
\title{Brief Description (1 parag) goes here}
\description{
Further description goes here, parags separated by empty
(Roxygen-block) lines
}
\author{
Ari Verbyla (Ari.Verbyla at csiro.au)
}

