% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mpwgaim.default.R
\name{mpwgaim.default}
\alias{mpwgaim.default}
\title{mpwgaim for method 'default'}
\usage{
\method{mpwgaim}{default}(baseModel, ...)
}
\arguments{
\item{baseModel}{a model object}

\item{\ldots}{Additional parameter}
}
\description{
Internal mpwgaim function
}
\details{
Default method for fitting mpwgaim.  The only available method is
'asreml'.  This function simply issues a message and stops.
}

