% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvmpwgaim.asreml.R
\name{mvmpwgaim.asreml}
\alias{mvmpwgaim.asreml}
\title{QTL analysis for multivariate multi-parent data}
\usage{
\method{mvmpwgaim}{asreml}(baseDiag, baseModel, phenoData, intervalObj,
  Trait = NULL, merge.by = NULL, gen.type = "interval", n.fa = 1,
  TypeI = 0.05, attempts = 10, data.name = NULL, trace = FALSE,
  verboseLev = 0, main.effects = FALSE, restart = NULL,
  dorestart = FALSE, ...)
}
\arguments{
\item{baseDiag}{A base model without marker based effects in
which all genetic effects associated with the multivariate nature
of the data are modelled using the 'diag' structure of 'asreml'
(see Details and Examples).}

\item{baseModel}{A base model without marker based effects in
which all genetic effects associated with the multivariate nature
of the data are modelled using the 'fa' structure of 'asreml'
(see Details and Examples).}

\item{phenoData}{Phenotypic data that includes all experimental
design and genetic line information that enables the base models
to be fitted.}

\item{intervalObj}{An object of class 'interval'.}

\item{Trait}{A character vector that specifies the name of the
multivariate trait or environment variable in the phenotypic data.}

\item{merge.by}{A character vector that specifies the genetic
line variable that is present in both the phenotypic data and the
'interval' object that is used in merging the two together.}

\item{gen.type}{A character vector indicating the type of
analysis to be undertaken.  Possible values are 'interval' (the
default) or 'marker'.}

\item{n.fa}{A number specifying the number of factors to be
included in the working model using founder probabilities.  The
default is 1 and the maximum is 2.  Note that the number can be
restricted by the number of traits or environments in the analysis.}

\item{TypeI}{The type I error rate for the forward selection of
putative QTL.  The default is 0.05.}

\item{attempts}{An integer representing the number of attempts at
convergence for the QTL model. The default is 5.}

\item{data.name}{A character vector specifying the name of the data
file to which the final data frame is to be saved.  If
\code{NULL}, the data frame is saved to the name of the response
dot data.}

\item{trace}{Logical for writing out debugging information.
Default is FALSE.}

\item{verboseLev}{An integer greater than 0 will result in
printing of outlier statistics in the QTL selection process.}

\item{main.effects}{Logical.  Should be \code{TRUE} for
multi-environment QTL analysis.  Otherwise the default is
\code{FALSE}.}

\item{restart}{A list with objects that allow the QTL analysis to
restart from the last successful QTL selection (see Details).}

\item{dorestart}{Logical.  Default is \code{FALSE}.  If
\code{TRUE}, the components required for a restart are saved after
each selection of a putative QTL.  This allows the process to be
restarted in the event that computational or time limits are
reached in the current analysis.}

\item{...}{Additional objects that can be passed to functions, for
example 'asreml'.}
}
\value{
A list object that has class 'mvmpwgaim' as well as
'asreml'.  There is an additional list, \code{QTL}, which has
many components with information on the multi-parent population,
arguments set, QTL and diagnostics.
}
\description{
Multivariate (multi-trait or multi-environment) QTL analysis for
multi-parent populations
}
\details{
This function allows either a multi-trait or multi-environment QTL
analysis to be carried out.  Note that the current code does not
allow multiple traits and multiple environments in the same
analysis.  Two base models are required.  In the first, the
polygenic effect MUST have a diagonal structure for the trait or
environment variable, while in the second the polygenic effect
MUST have a factor analytic structure for the trait or environment
variable.  The diagonal form is used to establish if there is
sufficient evidence to select a putative QTL while the factor
analytic form is used both for selection and computation of impact
of the QTL through probability measures, log probability measures
and percentage variance accounted for by each QTL.

For multi-environment analysis the logical argument main.effects
should be set to \code{TRUE}.  This ensures the possibility of
testing for environment by QTL interaction (see \code{remlrt}).

Analysis using this function is both time consuming and
computationally demanding.  To facilitate analyses that take along
time, there is a logical variable that can be set, dorestart.  If
this argument is set to \code{TRUE}, a list of structures is saved
to file after each selection of a QTL.  If the analysis terminates
prematurely, this saved file can be used as input in a subsequent
run of the analysis through the restart argument.  The name of the
file saved is \code{"restart.RData"} and this should be the
argument in a subsequent call to \code{mvmpwgaim}.
}
\author{
Ari Verbyla (ari.verbyla at csiro.au) and Klara Verbyla
(klara.verbyla at csiro.au)
}
\references{
Verbyla, A. P., Cavanagh, C. C. and Verbyla,
K. L. (2014). Whole genome analysis of multi-environment or
multi-trait QTL in MAGIC.  Genes, Genomes, Genetics.
}

