% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pchisq.mixture.R
\name{pchisq.mixture}
\alias{pchisq.mixture}
\title{Probabilities for specified quantiles of a mixture of
chi-square distributions}
\usage{
pchisq.mixture(x, ntrait = 2)
}
\arguments{
\item{x}{Quantiles of the mixture distribution.}

\item{ntrait}{Number of traits or environments (or the size in
the underlying binomial distribution)}
}
\value{
Probabilities for the specified quantiles are returned.
}
\description{
Probability distribution function for specified quantiles of a
mixture of chi-square distributions
}
\details{
Probabilities for given quantiles are found from a mixture of
chi-square distribution based on mixing probabilities from a
binomial distribution with size equal to the number of traits or
environments and probability of "success" equal to 0.5.
}
\examples{
## Simple example
pchisq.mixture(5, ntrait=4)
}
\author{
Ari Verbyla (ari.verbyla at csiro.au)
}

