% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.mpwgaim.R
\name{print.mpwgaim}
\alias{print.mpwgaim}
\title{Print method for 'mpwgaim'}
\usage{
\method{print}{mpwgaim}(x, intervalObj, ...)
}
\arguments{
\item{x}{an object of class 'mpwgaim'}

\item{intervalObj}{an object of class 'interval'}

\item{\ldots}{Additional optional objects}
}
\description{
Print method for mpwgaim objects
}
\details{
Provides a simple printing of the selected QTL after using
'mpwgaim'.
}
\author{
Julian Taylor (julian.taylor at adelaide.edu.au) and Ari
Verbyla (ari.verbyla at csiro.au)
}

