% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.summary.mpwgaim.R
\name{print.summary.mpwgaim}
\alias{print.summary.mpwgaim}
\title{Print table of QTL information}
\usage{
\method{print}{summary.mpwgaim}(x, ...)
}
\arguments{
\item{x}{an object of class 'summary.mpwgaim'}

\item{\ldots}{other arguments passed to print.summary.mpwgaim
(ignored)}
}
\description{
Print method for objects of class 'summary.mpwgaim'
}
\details{
Prints the summary information found using 'summary.mpwgaim'
}
\author{
Ari Verbyla (ari.verbyla at csiro.au)
}

