% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qchisq.mixture.R
\name{qchisq.mixture}
\alias{qchisq.mixture}
\title{Quantiles of a mixture of chi-square distributions}
\usage{
qchisq.mixture(prob, ntrait = 2, trace = F, maxiter = 10)
}
\arguments{
\item{prob}{A single or vector of probabilities (}

\item{ntrait}{Number of traits or environments (or the binomial
size)}

\item{trace}{Logical set to \code{FALSE}.  If \code{TRUE} then
the iterative process is printed.}

\item{maxiter}{Maximum number of iterations to carry out.  The
default is 10.}
}
\value{
The quantiles for the specified probabilities are returned.
}
\description{
Internal function for quantiles of a mixture of chi-square
distributions
}
\details{
An iterative process is used to determine the quantiles for
specified probabilities in \code{prob}, with mixing probabilities in
the mixture from a binomial distribution with size the number of
traits or environments (in the context of mvmpwgaim) and
probability of success equal to 0.5.  The process is stable for
probabilities that are not too small.  For a critical value in a
test using this distribution with type I error 0.05, the
probability should be specified as 0.95 (so that the probability is
assumed to relate to values of the mixture distribution less than
the critical value).
}
\examples{
## Simple examples
qchisq.mixture(0.95, ntrait=4)
qchisq.mixture(0.05, ntrait=4) ## fails
}
\author{
Ari Verbyla (ari.verbyla at csiro.au)
}

