% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/remlrt.default.R
\name{remlrt.default}
\alias{remlrt.default}
\title{Default method for remlrt}
\usage{
\method{remlrt}{default}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mvmpwgaim}}

\item{\ldots}{Optional additional arguments.}
}
\value{
Error message.
}
\description{
Default method for remlrt
}
\details{
Default method which returns an error message
}
\author{
Ari Verbyla (ari.verbyla at csiro.au)
}

